+++ stock-game/package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^12.8.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.0.1",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.1.2"
   },
