+++ stock-game/src/Profile.js
@@ -1,0 +1,14 @@
+import React from "react";
+import './components/StyleSheet.css'
+
+function Profile() {
+    return(
+        <>
+        <div>
+            Hello!
+        </div>
+        </>
+    )
+}
+
+export default Profile;
