+++ stock-game/src/components/StyleSheet.css
@@ -1,6 +1,17 @@
-.primary {
-    color: green;
-}
+body {
+    margin: 0;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+      'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+      sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+  }
+  
+  code {
+    font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+      monospace;
+  }
+  
 
 .App-header {
     background-color: #282c34;
@@ -12,6 +23,12 @@
     font-size: calc(5px + 2vmin);
     color: white;
   }
+
+.Page-Default {
+
+    margin: 0;
+
+}
 
 .button {
 background-color: #52B788; /* Green */
